requirements.txt
Flask
pymongo
python-dotenv
google-generativeai
langchain
PyPDF2
faiss-cpu
langchain-google-genai 